requirements.txt
pillow
python-resize-image
wget
